Move graphqlHTTP into own file
